rename juridische-invloed-tooptip
renamed file and all imports
