+++ pages/houtbouw/gemeentelijke-omgevingsvisie.js
@@ -4,7 +4,7 @@ import Layout from '../../components/layout';
 import IcontWood from '../../public/icons/houtbouwIconBg.svg';
 import RTooltip from '../../components/r_ladder_tooltip';
 import JHTooltip from '../../components/juridische-houdbaarheid-tooltip';
-import JITooltip from '../../components/juridische_invloed_tooltip';
+import JITooltip from '../../components/juridische-invloed-tooltip';
 
 function classNames(...classes) {
   return classes.filter(Boolean).join(' ');
