+++ pages/measures/[id].jsx
@@ -7,7 +7,7 @@ import WindmillIcon from '../../public/winturbines.svg';
 import IcontWood from '../../public/icons/houtbouwIconBg.svg';
 import RTooltip from '../../components/r_ladder_tooltip';
 import JHTooltip from '../../components/juridische-houdbaarheid-tooltip';
-import JITooltip from '../../components/juridische_invloed_tooltip';
+import JITooltip from '../../components/juridische-invloed-tooltip';
 
 const formatDate = (date) => {
   let dateObject = new Date(date);
